+++ lib/onlyoffice_testrail_wrapper/testrail_api_object.rb
@@ -10,5 +10,17 @@ module OnlyofficeTestrailWrapper
       hash.each { |key, value| instance_variable_set("@#{key}", value) }
       self
     end
+
+    # Generate hash wih pars of name:id
+    # @param [Array<Hash>] array of raw data
+    # @param [String, Symbol] name_key how name key is called in hash
+    # @return [Hash] result hash data
+    def name_id_pairs(array, name_key = 'name')
+      raise 'First argument must be Array!' unless array.is_a?(Array)
+
+      result_hash = {}
+      array.reverse_each { |element| result_hash[element[name_key]] = element['id'] }
+      result_hash
+    end
   end
 end