+++ lib/onlyoffice_testrail_wrapper/testrail_project/testrail_project_plan_helper.rb
@@ -54,7 +54,7 @@ module OnlyofficeTestrailWrapper
       get_url = "index.php?/api/v2/get_plans/#{@id}"
       filters.each { |key, value| get_url += "&#{key}=#{value}" }
       plans = Testrail2.http_get(get_url)
-      @plans_names = HashHelper.get_hash_from_array_with_two_parameters(plans, 'name', 'id') if @plans_names.empty?
+      @plans_names = name_id_pairs(plans) if @plans_names.empty?
       plans
     end