+++ lib/onlyoffice_testrail_wrapper/testrail_project/testrail_project_suite_methods.rb
@@ -18,7 +18,7 @@ module OnlyofficeTestrailWrapper
     # @return [Array<Hash>] array with suites data in hash
     def get_suites
       suites = Testrail2.http_get("index.php?/api/v2/get_suites/#{@id}")
-      @suites_names = HashHelper.get_hash_from_array_with_two_parameters(suites, 'name', 'id') if @suites_names.empty?
+      @suites_names = name_id_pairs(suites) if @suites_names.empty?
       suites
     end